requirements.txt
Flask
Flask-Bootstrap
Flask-Login
werkzeug
python-dotenv
psycopg2-binary